Trimly, Criando regras de negócio para a roleService

+++ src/services/RolesService.ts
@@ -19,7 +19,19 @@ export class RolesService {
     async addUserRoles(userId: number, params: RoleTypeAttributes) {
         const userExists = await this.userRepository.findById(userId)
         if (!userExists) throw new HttpError(404, "User not found");
-        const role = await this.rolesRepository.findByRoleType(params.role_type)
+        const userRole = await this.rolesRepository.findByUserIdRoles(userId)
+        if (!userRole) throw new HttpError(404, "User not Role");
+        const userAdmin = userRole.find(element => {
+            if (element.roles.role_type === "Admin") {return true} else {return false}
+        });
+        let role 
+        if (userAdmin){
+            role = await this.rolesRepository.findByRoleType(params.role_type)
+        } else if (params.role_type !== "Admin"){
+            role = await this.rolesRepository.findByRoleType(params.role_type)
+        } else {
+            throw new HttpError(403, "permission denied!");
+        }
         const roleId = role?.id;
         if (!roleId) {
             throw new HttpError(404, "Role not found");
@@ -37,7 +49,20 @@ export class RolesService {
         if (!userExists) throw new HttpError(404, "User not found");
         const roleExists = await this.rolesRepository.findByRoleId(roleId)
         if (!roleExists) throw new HttpError(404, "Role not found");
-        const role = await this.rolesRepository.findByRoleType(params.role_type)
+        if (roleExists.role_type === "Client") throw new HttpError(403, "Unable to change client role!")
+        const userRole = await this.rolesRepository.findByUserIdRoles(userId)
+        if (!userRole) throw new HttpError(404, "User not Role");
+        const userAdmin = userRole.find(element => {
+            if (element.roles.role_type === "Admin") {return true} else {return false}
+        });
+        let role 
+        if (userAdmin){
+            role = await this.rolesRepository.findByRoleType(params.role_type)
+        } else if (params.role_type !== "Admin"){
+            role = await this.rolesRepository.findByRoleType(params.role_type)
+        } else {
+            throw new HttpError(403, "permission denied!");
+        }
         const newRoleId = role?.id;
         if (!newRoleId) {
             throw new HttpError(404, "New Role not found");
@@ -59,7 +84,12 @@ export class RolesService {
         if (!userExists) throw new HttpError(404, "User not found");
         const roleExists = await this.rolesRepository.findByUserRoleExist(userId, roleId);
         if (!roleExists) throw new HttpError(404, "Role not found");
-        const deletedRole = await this.rolesRepository.deletedByUserIdRoleId(userId, roleId)
-        return deletedRole
+        const roleExistsName = await this.rolesRepository.findByRoleId(roleId)
+        if (roleExistsName?.role_type !== "Client") {
+            const deletedRole = await this.rolesRepository.deletedByUserIdRoleId(userId, roleId)
+            return deletedRole
+        } else{
+            throw new HttpError(403, "Unable to delete client role!")
+        }
     }
 }
